Custom Myelin kernels, dynamic instance allocation, etc.

+++ myelin/kernel/generic-matmul.cc
@@ -134,8 +134,8 @@ class GenericFltVecMatMulBase : public Kernel {
     __ j(not_equal, &l1);
   }
 
-  int Complexity(const Step *step) override {
-    int ops = step->input(1)->elements() * 2;
+  int64 Complexity(const Step *step) override {
+    int64 ops = step->input(1)->elements() * 2;
     if (bias_) ops += step->input(2)->elements();
     if (relu_) ops += step->output(0)->elements();
     return ops;
@@ -297,7 +297,7 @@ class GenericFltMatMatMul : public Kernel {
     __ j(not_equal, &l1);
   }
 
-  int Complexity(const Step *step) override {
+  int64 Complexity(const Step *step) override {
     return step->input(0)->dim(0) * step->input(1)->elements() * 2;
   }
 };
@@ -453,7 +453,7 @@ class GenericIntVecMatMulBase : public Kernel {
     __ j(not_equal, &l1);
   }
 
-  int Complexity(const Step *step) override {
+  int64 Complexity(const Step *step) override {
     int ops = step->input(1)->elements() * 2;
     if (bias_) ops += step->input(2)->elements();
     if (relu_) ops += step->output(0)->elements();
